Fix trivially fixable warnings (#411)
3/2 more serious warnigns on gcc/clang remain.

A bunch of -Wdeprecated-copy warnings have been fixed
E.g. class Point had explicit Copy C-tor which forces an implicit copy
asignment operator. -Wdeprecated-copy complains about this. See rule of
five.

Several signed/unsigned comparisons could be removed by changing loop
variables to size_t instead of int or unsigned int.

Missing return statements added several functions.

Several variable shadowing warnings have been fixed by renaming the
shadowing variable.

Unused parameters have been omited from function signatures so that no
UNUSED macro is neccesary anymore. Also the macro has been removed since
the lambda capture trick now triggers a warning on clang for an unused
cpature. Since now all unused vars are gone I simply removed the macro.

Some unneccesary const casts have been removed.

+++ Simulation.cpp
@@ -361,7 +361,7 @@ void Simulation::UpdateRoutesAndLocations()
 //    }
 
 #pragma omp parallel for shared(pedsToRemove, allRooms)
-     for (signed int p = 0; p < allPeds.size(); ++p) {
+     for (size_t p = 0; p < allPeds.size(); ++p) {
           auto ped = allPeds[p];
           Room* room = _building->GetRoom(ped->GetRoomID());
           SubRoom* sub0 = room->GetSubRoom(ped->GetSubRoomID());
